Update project_recursion.md
recursively generate an array of the Fibonacci sequence.

after going and reading the given blog. I did not find a good solution to the given problem. it was briefly mentioned and done outside the function. So I linked a link to a post with the exact problem. so people can compare it with their code and look at it if they are stuck. consdering the lesson is from a different diffuclty than the previous lessons.